requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-paypal==2.0
idna==3.6
pillow==10.3.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
stripe==8.10.0
typing_extensions==4.11.0
urllib3==2.2.1
